*** empty log message ***

+++ coders/pict.c
@@ -1503,9 +1503,9 @@ static unsigned int WritePICTImage(const ImageInfo *image_info,Image *image)
   /*
     Write full size of the file, resolution, frame bounding box, and reserved.
   */
-  WriteBlobMSBShort(image,image->x_resolution);
+  WriteBlobMSBShort(image,(unsigned short) image->x_resolution);
   WriteBlobMSBShort(image,0x0000);
-  WriteBlobMSBShort(image,image->y_resolution);
+  WriteBlobMSBShort(image,(unsigned short) image->y_resolution);
   WriteBlobMSBShort(image,0x0000);
   WriteBlobMSBShort(image,frame_rectangle.top);
   WriteBlobMSBShort(image,frame_rectangle.left);
@@ -1587,9 +1587,9 @@ static unsigned int WritePICTImage(const ImageInfo *image_info,Image *image)
       WriteBlobMSBShort(image,768);
       WriteBlobMSBShort(image,image->columns);
       WriteBlobMSBShort(image,image->rows);
-      WriteBlobMSBShort(image,image->x_resolution);
+      WriteBlobMSBShort(image,(unsigned short) image->x_resolution);
       WriteBlobMSBShort(image,0x0000);
-      WriteBlobMSBShort(image,image->y_resolution);
+      WriteBlobMSBShort(image,(unsigned short) image->y_resolution);
       WriteBlobMSBLong(image,0x00000000UL);
       WriteBlobMSBLong(image,0x82B60001UL);
       WriteBlobMSBLong(image,0x0C50685FUL);
@@ -1626,9 +1626,9 @@ static unsigned int WritePICTImage(const ImageInfo *image_info,Image *image)
   WriteBlobMSBShort(image,pixmap.version);
   WriteBlobMSBShort(image,pixmap.pack_type);
   WriteBlobMSBLong(image,pixmap.pack_size);
-  WriteBlobMSBShort(image,image->x_resolution);
+  WriteBlobMSBShort(image,(unsigned short) image->x_resolution);
   WriteBlobMSBShort(image,0x0000);
-  WriteBlobMSBShort(image,image->y_resolution);
+  WriteBlobMSBShort(image,(unsigned short) image->y_resolution);
   WriteBlobMSBShort(image,0x0000);
   WriteBlobMSBShort(image,pixmap.pixel_type);
   WriteBlobMSBShort(image,pixmap.bits_per_pixel);
